Rename internal Provider to Web3Provider
